requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
requests
tkcalendar
openpyxl
